requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
Flask-Caching==1.8.0
PyMySQL==0.9.3
simplejson==3.17.0
dependency-injector==3.15.6
PyYAML==5.3

# Test
pytest==5.3.5
pycodestyle==2.5.0
docker==4.2.0
